sync: move sync code from santactl dir to santasyncservice dir
