+++ Source/santasyncservice/SNTCommandSyncPreflight.m
@@ -12,7 +12,7 @@
 ///    See the License for the specific language governing permissions and
 ///    limitations under the License.
 
-#import "Source/santactl/Commands/sync/SNTCommandSyncPreflight.h"
+#import "Source/santasyncservice/SNTCommandSyncPreflight.h"
 
 #import <MOLXPCConnection/MOLXPCConnection.h>
 
@@ -21,8 +21,8 @@
 #import "Source/common/SNTLogging.h"
 #import "Source/common/SNTSystemInfo.h"
 #import "Source/common/SNTXPCControlInterface.h"
-#import "Source/santactl/Commands/sync/SNTCommandSyncConstants.h"
-#import "Source/santactl/Commands/sync/SNTCommandSyncState.h"
+#import "Source/santasyncservice/SNTCommandSyncConstants.h"
+#import "Source/santasyncservice/SNTCommandSyncState.h"
 
 @implementation SNTCommandSyncPreflight
