+++ Source/santasyncservice/SNTCommandSyncState.h
