+++ Source/santasyncservice/testdata/sync_eventupload_input_basic.plist
