+++ Source/santasyncservice/testdata/sync_ruledownload_batch1.json
