+++ Source/santasyncservice/testdata/sync_ruledownload_batch2.json
